refactor/clean-code-for-server-send
